first version of lib-autopilot.py
the lib can currently generate candidates for an autopilot to open channels with according to four strategies:
Random: following the Erdoes Renyi model nodes are drawn from a uniform distribution
Central: nodes are sampled from a uniform distribution of the top most central nodes (betweenness)
Network_Improovement: nodes are sampled from a uniform distribution of the nodes which are badly connected
richness: nodes with high liquidity are taken and it is sampled from a uniform distribution of those

this library is work in progress. there is no logic yet implemented on how many funds to allocate to which channel.
also due to the randomness the lib might sometimes find less channels than was proposed.
also many conor cases will let the code break.